requirements.txt
Django==3.2.12
psycopg2==2.9.3
djangorestframework==3.13.1
xmltodict==0.12.0
requests==2.27.1